+++ hooks/useSelectedPriceCurrency.ts
@@ -1,0 +1,23 @@
+import BigNumber from 'bignumber.js';
+import useExchangeRatesQuery from 'queries/rates/useExchangeRatesQuery';
+
+import { useRecoilValue } from 'recoil';
+import { priceCurrencyState } from 'store/app';
+
+const useSelectedPriceCurrency = () => {
+	const selectedPriceCurrency = useRecoilValue(priceCurrencyState);
+	const exchangeRatesQuery = useExchangeRatesQuery();
+	const exchangeRates = exchangeRatesQuery.data ?? null;
+	const selectPriceCurrencyRate = exchangeRates && exchangeRates[selectedPriceCurrency.name];
+
+	const getPriceAtCurrentRate = (price: BigNumber) =>
+		selectPriceCurrencyRate != null ? price.dividedBy(selectPriceCurrencyRate) : price;
+
+	return {
+		selectPriceCurrencyRate,
+		selectedPriceCurrency,
+		getPriceAtCurrentRate,
+	};
+};
+
+export default useSelectedPriceCurrency;