+++ sections/dashboard/TrendingSynths/SynthRow.tsx
@@ -15,26 +15,23 @@ import useHistoricalRatesQuery from 'queries/rates/useHistoricalRatesQuery';
 import useHistoricalVolumeQuery from 'queries/rates/useHistoricalVolumeQuery';
 
 import { SelectableCurrencyRow } from 'styles/common';
+import useMarketClosed from 'hooks/useMarketClosed';
+import useSelectedPriceCurrency from 'hooks/useSelectedPriceCurrency';
 
 type SynthRowProps = {
 	price: number | null;
 	synth: Synth;
-	selectedPriceCurrency: Synth;
-	selectPriceCurrencyRate: number | null;
 };
-const SynthRow: FC<SynthRowProps> = ({
-	price,
-	synth,
-	selectedPriceCurrency,
-	selectPriceCurrencyRate,
-}) => {
+const SynthRow: FC<SynthRowProps> = ({ price, synth }) => {
 	const { t } = useTranslation();
 	const router = useRouter();
+	const { selectPriceCurrencyRate, selectedPriceCurrency } = useSelectedPriceCurrency();
 
 	const currencyKey = synth.name;
 
 	useHistoricalVolumeQuery(currencyKey, Period.ONE_DAY);
 	const historicalRates = useHistoricalRatesQuery(currencyKey, Period.ONE_DAY);
+	const { marketClosureReason } = useMarketClosed(currencyKey);
 
 	return (
 		<StyledSelectableCurrencyRow
@@ -47,6 +44,7 @@ const SynthRow: FC<SynthRowProps> = ({
 					currencyName: synth.description,
 				})}
 				showIcon={true}
+				marketClosureReason={marketClosureReason}
 			/>
 			{price != null ? (
 				<Currency.Price