+++ sections/homepage/Benefits/Benefits.tsx
@@ -3,8 +3,10 @@ import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 import Img from 'react-optimized-image';
 
-import SwapPreview from 'assets/svg/marketing/swap-preview.svg';
-import SwapPreviewMd from 'assets/svg/marketing/swap-preview-md.svg';
+import { SYNTHS_MAP } from 'constants/currency';
+
+// import SwapPreview from 'assets/svg/marketing/swap-preview.svg';
+// import SwapPreviewMd from 'assets/svg/marketing/swap-preview-md.svg';
 import P2CIcon from 'assets/svg/marketing/icon-p2c.svg';
 import PermissionlessIcon from 'assets/svg/marketing/icon-permissionless.svg';
 import RestrictionsIcon from 'assets/svg/marketing/icon-restrictions.svg';
@@ -14,7 +16,7 @@ import { FlexDivCol } from 'styles/common';
 import media from 'styles/media';
 
 import { StackSection, CenterSubHeader, Title, Copy } from '../common';
-import { DesktopOnlyView, MobileOrTabletView } from 'components/Media';
+import useExchange from 'sections/exchange/hooks/useExchange';
 
 const BENEFITS = [
 	{
@@ -39,16 +41,21 @@ const BENEFITS = [
 
 const Benefits = () => {
 	const { t } = useTranslation();
+	const { quoteCurrencyCard, baseCurrencyCard, footerCard } = useExchange({
+		footerCardAttached: true,
+		persistSelectedCurrencies: false,
+		defaultQuoteCurrencyKey: SYNTHS_MAP.sBTC,
+		defaultBaseCurrencyKey: SYNTHS_MAP.sETH,
+	});
 
 	return (
 		<StackSection>
 			<StyledCenterSubHeader>{t('homepage.benefits.title')}</StyledCenterSubHeader>
-			<DesktopOnlyView>
-				<ResponsiveImage src={SwapPreview} alt="" webp={true} />
-			</DesktopOnlyView>
-			<MobileOrTabletView>
-				<ResponsiveImage src={SwapPreviewMd} alt="" webp={true} />
-			</MobileOrTabletView>
+			<ExchangeCards>
+				{quoteCurrencyCard}
+				{baseCurrencyCard}
+			</ExchangeCards>
+			<ExchangeFooter>{footerCard}</ExchangeFooter>
 			<BenefitContainer>
 				{BENEFITS.map(({ id, image, title, copy }) => (
 					<BenefitCard key={id}>
@@ -61,6 +68,47 @@ const Benefits = () => {
 		</StackSection>
 	);
 };
+
+const ExchangeFooter = styled.div`
+	width: 100%;
+	.footer-card {
+		padding: 16px 18px;
+		width: 800px;
+		max-width: unset;
+	}
+	${media.lessThan('md')`
+		.footer-card {
+			width: unset;
+			position: unset;
+			box-shadow: unset;
+		}
+		.trade-summary-card {
+			margin-bottom: 24px;
+		}
+	`}
+`;
+
+const ExchangeCards = styled.div`
+	width: 100%;
+	display: grid;
+	justify-content: center;
+	grid-template-columns: auto auto;
+	grid-gap: 2px;
+	padding-bottom: 2px;
+
+	.currency-card {
+		width: 400px;
+	}
+	${media.lessThan('md')`
+		justify-content: unset;
+		grid-template-columns: unset;
+		grid-template-rows: auto auto;
+		.currency-card {
+			width: 100%;
+		}
+		padding-bottom: 24px;
+	`}
+`;
 
 export const ResponsiveImage = styled(Img)`
 	width: 100%;