+++ sections/shared/modals/SelectQuoteCurrencyModal/SynthRow.tsx
@@ -2,53 +2,50 @@ import { FC } from 'react';
 import styled from 'styled-components';
 import { useTranslation } from 'react-i18next';
 
-import { Synth } from 'lib/synthetix';
+import synthetix from 'lib/synthetix';
 
 import Currency from 'components/Currency';
 
-import { NO_VALUE } from 'constants/placeholder';
+import useMarketClosed from 'hooks/useMarketClosed';
+
+import { SynthBalance } from 'queries/walletBalances/useSynthsBalancesQuery';
 
 import { SelectableCurrencyRow } from 'styles/common';
-import useHistoricalRatesQuery from 'queries/rates/useHistoricalRatesQuery';
-import { Period } from 'constants/period';
-import useMarketClosed from 'hooks/useMarketClosed';
 import useSelectedPriceCurrency from 'hooks/useSelectedPriceCurrency';
 
 type SynthRowProps = {
-	price: number | null;
-	synth: Synth;
+	synth: SynthBalance;
 	onClick: () => void;
 };
-const SynthRow: FC<SynthRowProps> = ({ price, synth, onClick }) => {
+const SynthRow: FC<SynthRowProps> = ({ synth, onClick }) => {
 	const { t } = useTranslation();
 	const { selectPriceCurrencyRate, selectedPriceCurrency } = useSelectedPriceCurrency();
 
-	const currencyKey = synth.name;
+	const { synthsMap } = synthetix;
+	const { currencyKey, usdBalance, balance } = synth;
+	const synthDesc = synthsMap != null ? synthsMap[currencyKey]?.description : null;
 
-	const historicalRates = useHistoricalRatesQuery(currencyKey, Period.ONE_DAY);
+	const totalValue = usdBalance;
+
 	const { marketClosureReason } = useMarketClosed(currencyKey);
 
 	return (
-		<StyledSelectableCurrencyRow key={currencyKey} onClick={onClick} isSelectable={true}>
+		<StyledSelectableCurrencyRow onClick={onClick} isSelectable={true}>
 			<Currency.Name
 				currencyKey={currencyKey}
 				name={t('common.currency.synthetic-currency-name', {
-					currencyName: synth.description,
+					currencyName: synthDesc,
 				})}
 				showIcon={true}
 				marketClosureReason={marketClosureReason}
 			/>
-			{price != null ? (
-				<Currency.Price
-					currencyKey={currencyKey}
-					price={price}
-					sign={selectedPriceCurrency.sign}
-					conversionRate={selectPriceCurrencyRate}
-					change={historicalRates.data?.change}
-				/>
-			) : (
-				NO_VALUE
-			)}
+			<Currency.Amount
+				currencyKey={currencyKey}
+				amount={balance}
+				totalValue={totalValue}
+				sign={selectedPriceCurrency.sign}
+				conversionRate={selectPriceCurrencyRate}
+			/>
 		</StyledSelectableCurrencyRow>
 	);
 };