+++ utils/network.ts
@@ -30,3 +30,5 @@ export const getTransactionPrice = (
 };
 
 export const normalizeGasLimit = (gasLimit: number) => gasLimit + DEFAULT_GAS_BUFFER;
+
+export const gasPriceInWei = (gasPrice: number) => gasPrice * GWEI_UNIT;